% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeomFlatViolin.R
\docType{data}
\name{GeomFlatViolin}
\alias{GeomFlatViolin}
\title{Helper function to plot half-violins in ggplots}
\format{
An object of class \code{GeomFlatViolin} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.
}
\usage{
GeomFlatViolin
}
\description{
Code from David Robinson
https://gist.github.com/dgrtwo/eb7750e74997891d7c20
}
\examples{

}
\keyword{flat;}
\keyword{violin}
