% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_flat_violin.R
\name{\%||\%}
\alias{\%||\%}
\title{Helper function to for flat violin geoms}
\usage{
a \%||\% b
}
\description{
Code from David Robinson
https://gist.github.com/dgrtwo/eb7750e74997891d7c20
}
\examples{

}
\keyword{flat;}
\keyword{violin}
